move functionality of cluster_one_assembly to cluster_multiple_assemblies and delete the former
